Export CollectionSubscription type
